Obtener bucket y requirement por ID

+++ src/main/java/com/example/demo/dao/IBucketDAO.java
@@ -7,6 +7,7 @@ import com.example.demo.model.Bucket;
 
 
 @Repository
-public interface IBucketDAO extends JpaRepository<Bucket, Long>{
+public interface IBucketDAO extends JpaRepository<Bucket, Integer>{
+
 
 }
